Extract node resolver into a separate library
