+++ packages/resolvers/default/package.json
@@ -16,9 +16,7 @@
   },
   "dependencies": {
     "@parcel/plugin": "^2.0.0-alpha.3.1",
-    "@parcel/utils": "^2.0.0-alpha.3.1",
-    "micromatch": "^3.0.4",
-    "node-libs-browser": "^2.1.0"
+    "@parcel/node-resolver-core": "^2.0.0-alpha.3.1"
   },
   "devDependencies": {
     "@babel/core": "^7.2.2"
